perf(task-coordinator): split headless plugin entry
